389 fe edit my resources (#425)
* update editing resources, adding EditResource component and reuse it

* editResources and tests

* updated with changes requested

* fix code smell on CardResource
