+++ front/src/__tests__/organisms/ResourceForm.test.tsx
@@ -1,6 +1,6 @@
 import { vi } from 'vitest'
 import { fireEvent, render, screen, waitFor } from '../test-utils'
-import { ResourceForm } from '../../components/organisms'
+import ResourceForm from '../../components/organisms/ResourceForm'
 import { paths } from '../../constants'
 
 const navigate = vi.fn()
@@ -140,7 +140,7 @@ describe('ResourceForm', () => {
       screen.getByDisplayValue('Renderizado condicional')
     ).toBeInTheDocument()
 
-    const button = screen.getByText(/actualizar/i)
+    const button = screen.getByText(/editar/i)
     fireEvent.click(button)
     await waitFor(() => {
       expect(navigate).toHaveBeenCalledWith(paths.home)
@@ -169,7 +169,7 @@ describe('ResourceForm', () => {
     fireEvent.change(screen.getByLabelText(/título/i), {
       target: { value: '' },
     })
-    fireEvent.click(screen.getByText(/actualizar/i))
+    fireEvent.click(screen.getByText(/editar/i))
     await waitFor(() =>
       expect(screen.getByText('Este campo es obligatorio')).toBeInTheDocument()
     )
