+++ front/src/components/organisms/ResourceCardList.tsx
@@ -4,8 +4,7 @@ import { FC } from 'react'
 import { urls } from '../../constants'
 import { FlexBox, dimensions } from '../../styles'
 import { Spinner, Text } from '../atoms'
-// eslint-disable-next-line import/no-cycle
-import { CardResource } from '../molecules'
+import CardResource from './CardResource'
 import { TFilters, buildQueryString } from '../../helpers'
 import { useSortByDate } from '../../hooks/useSortByDate'
 import { useAuth } from '../../context/AuthProvider'
