+++ front/src/components/organisms/ResourceForm.tsx
@@ -4,9 +4,7 @@ import { zodResolver } from '@hookform/resolvers/zod'
 import { useNavigate } from 'react-router-dom'
 import { useMutation } from '@tanstack/react-query'
 import styled from 'styled-components'
-// eslint-disable-next-line import/no-cycle
 import { ChangeEvent, FC } from 'react'
-// eslint-disable-next-line import/no-cycle
 import { InputGroup, SelectGroup } from '../molecules'
 import { Button, ValidationMessage, Radio } from '../atoms'
 import { FlexBox, dimensions } from '../../styles'
@@ -46,11 +44,6 @@ const ResourceFormSchema = z.object({
   userEmail: z.string().optional(),
 })
 
-// export type TResourceForm = z.infer<typeof ResourceFormSchema> & {
-//   topics: string[]
-//   topicId?: string
-//   id?: string
-// }
 export type TResourceForm = Omit<
   z.infer<typeof ResourceFormSchema>,
   'topics'
@@ -110,7 +103,7 @@ const updateResourceFetcher = (resource: object) =>
     // eslint-disable-next-line no-console
     .catch((error) => console.error(error))
 
-export const ResourceForm: FC<TSelectOptions> = ({
+const ResourceForm: FC<TSelectOptions> = ({
   selectOptions,
   initialValues,
   resourceId,
@@ -138,7 +131,7 @@ export const ResourceForm: FC<TSelectOptions> = ({
       navigate(paths.home)
     },
   })
-  const onSubmit = handleSubmit(async (data) => {
+  const create = handleSubmit(async (data) => {
     const { title, description, url, topics, resourceType } = data
 
     await createResource.mutateAsync({
@@ -149,7 +142,7 @@ export const ResourceForm: FC<TSelectOptions> = ({
       resourceType,
     })
   })
-  const onSubmitUpdate = handleSubmit(async (data) => {
+  const update = handleSubmit(async (data) => {
     const { title, description, url, topicId, resourceType } = data
 
     const updatedData = {
@@ -167,8 +160,12 @@ export const ResourceForm: FC<TSelectOptions> = ({
     const selectedTopicId = event.target.value
     setValue('topicId', selectedTopicId)
   }
+  // console.log(initialValues, 'initialValues')
   return (
-    <ResourceFormStyled onSubmit={initialValues ? onSubmitUpdate : onSubmit}>
+    <ResourceFormStyled
+      onSubmit={initialValues ? update : create}
+      data-testid="resource-form"
+    >
       <InputGroup
         hiddenLabel
         id="title"
@@ -227,8 +224,9 @@ export const ResourceForm: FC<TSelectOptions> = ({
         ) : null}
       </FlexErrorStyled>
       <ButtonContainerStyled align="stretch">
-        <Button type="submit">{initialValues ? 'Actualizar' : 'Crear'}</Button>
+        <Button type="submit">{initialValues ? 'Editar' : 'Crear'}</Button>
       </ButtonContainerStyled>
     </ResourceFormStyled>
   )
 }
+export default ResourceForm
