+++ front/src/components/organisms/index.ts
@@ -1,15 +1,15 @@
-// eslint-disable-next-line import/no-cycle
 import { CategoriesList } from './CategoriesList'
 import { HowToHelpCard } from './HowToHelpCard'
-// eslint-disable-next-line import/no-cycle
 import { ResourceCardList, TResource } from './ResourceCardList'
-import { ResourceForm, TResourceForm } from './ResourceForm'
+import type  {  TResourceForm } from './ResourceForm'
 import { TopicsRadioWidget } from './TopicsRadioWidget'
 import { Navbar } from './Navbar'
 import { MyFavoritesList } from './MyFavoritesList'
 import { MyResources } from './MyResources'
 import Login from './Login'
 import Register from './Register'
+import { CardResourceLink } from './CardResourceLink'
+import type { TCardResource } from './CardResource'
 
 export {
   ResourceCardList,
@@ -17,10 +17,10 @@ export {
   MyResources,
   HowToHelpCard,
   CategoriesList,
-  ResourceForm,
   Navbar,
   MyFavoritesList,
   Login,
   Register,
+  CardResourceLink
 }
-export type { TResourceForm, TResource }
+export type { TResourceForm, TResource, TCardResource}
